+++ src/App.js
@@ -26,7 +26,7 @@ const NotFound = lazy(() => import('pages/NotFound'))
 const Dashboard = lazy(() => import('pages/Dashboard'))
 const Search = lazy(() => import('pages/Search'))
 const ProjectSettings = lazy(() => import('pages/Extension/Settings'))
-const ExtensionPage = lazy(() => import('pages/Extensions/ExtensionPage'))
+const ExtensionPage = lazy(() => import('pages/Extension/View/ExtensionPage'))
 const minimalFooterPages = [
   '/projects', '/dashboard', '/settings', '/contact',
 ]
