+++ src/pages/Extension/Settings/ExtensionSettings.jsx
@@ -10,7 +10,6 @@ import _size from 'lodash/size'
 import _replace from 'lodash/replace'
 import _find from 'lodash/find'
 import _keys from 'lodash/keys'
-import _map from 'lodash/map'
 import _toNumber from 'lodash/toNumber'
 import _filter from 'lodash/filter'
 import _forEach from 'lodash/forEach'
